fix(CreateWorkoutModal): Prevent useEffect function to be executed for all modals
The issue was that when you add a workout through the creation workout modal on edit program page, you actually listen from all modals (on each days) for the store status change. Meaning that all the toaster was trigger for one change.

Solution is to return is the `isOpen` props is false, then we execute the useEffect function only for the current open modal